requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openpyxl==3.1.2
python-multipart==0.0.9
PyJWT==2.8.0
python-dotenv==1.0.1
